+++ src/main.ts
@@ -3,8 +3,7 @@ import Map from "ol/Map";
 import TileLayer from "ol/layer/WebGLTile";
 import View from "ol/View";
 import { Extent } from "ol/extent";
-import { colorPixel } from "./color";
-import { ExpressionValue } from "ol/style/webgl";
+import { colorPixelExpression } from "./colorGL";
 
 type MapCoords = [number, number];
 type ZoomCoords = [number, MapCoords];
@@ -61,15 +60,6 @@ const loadTile = (z: number, x: number, y: number): Promise<Uint8Array> => {
   });
 };
 
-function unpackUint32(data: number[]): number {
-  const byte1 = data[0] << 24;
-  const byte2 = data[1] << 16;
-  const byte3 = data[2] << 8;
-  const byte4 = data[3];
-
-  return byte1 | byte2 | byte3 | byte4;
-}
-
 const extent: Extent = [-80000000, -40000000, 60000000, 40000000];
 
 const view = new View({
@@ -81,49 +71,6 @@ const view = new View({
   center,
   zoom,
 });
-
-const colorPixelExpression = (): ExpressionValue => {
-  const normalizedIters = [
-    "floor",
-    ["+", ["*", ["band", 1], ["^", 2, 24]],
-    ["+", ["*", ["band", 2], ["^", 2, 16]],
-    ["+", ["*", ["band", 3], ["^", 2, 8]],
-    ["band", 4]]]]
-  ];
-  const value = ["*", normalizedIters, 255];
-  // const red = ["/", ["*", ["%", value, 8], 32], 255];
-  // const green = ["/", ["*", ["%", value, 16], 16], 255];
-  // const blue = ["/", ["*", ["%", value, 32], 8], 255];
-  const red = ["*", ["%", value, 8], 32];
-  const green = ["*", ["%", value, 16], 16];
-  const blue = ["*", ["%", value, 32], 8];
-  return ["color", red, green, blue, 1];
-  // return ["array", ["band", 1], ["band", 2], ["band", 3], ["band", 4]];
-};
-
-// const PALETTE_SCALE = 64;
-// const PALETTE_OFFSET = 0;
-// const colorPixelExpression = () => {
-//   // unpack Uint32 from Uint8Array bands
-//   const normalizedIters = [
-//     "floor",
-//     ["+", ["*", ["band", 1], ["^", 2, 24]]],
-//     ["+", ["*", ["band", 2], ["^", 2, 16]]],
-//     ["+", ["*", ["band", 3], ["^", 2, 8]]],
-//     ["+", ["band", 4]],
-//   ];
-
-//   const colorAdjust = [
-//     "+",
-//     ["*", PALETTE_SCALE, normalizedIters],
-//     PALETTE_OFFSET,
-//   ];
-//   const hue = ["%", [["*", normalizedIters], 360], []];
-//   const variance = ["+", 0.42, ["sin", ["*", 0.1]]];
-//   const blackPixel = ["array", 0, 0, 0, 1];
-
-//   return ["array", ["band", 1], ["band", 2], ["band", 3], ["band", 4]];
-// };
 
 const layer = new TileLayer({
   style: {
